fix type data lat and long

+++ models/auth.js
@@ -45,8 +45,8 @@ const SchoolAccount = sequelize.define('SchoolAccount', {
     type: DataTypes.STRING(500),
     allowNull: false,
   },
-  latitude: { type: DataTypes.FLOAT, allowNull: true },
-  longitude: { type: DataTypes.FLOAT, allowNull: true },
+  latitude: { type: DataTypes.DECIMAL(10, 8), allowNull: true },
+  longitude: { type: DataTypes.DECIMAL(11, 8), allowNull: true },
   email: {
     type: DataTypes.STRING(100),
     allowNull: false,
